# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Cadrly API
  version: 0.1.0
  description: Приложение для обмена коллекционными карточками
servers:
  - description: Test server
    url: http://localhost:8080/api
...
paths:
...
  /trades:
    get:
      tags: [Trades]
      summary: Получить все доступные предложения обмена
      description: |
        Возвращает список всех доступных предложений обмена.
        Можно провести поиск по названию карты.
      parameters:
        - in: query
          name: search
          description: Поиск по названию карты
          schema:
            type: string
      responses:
        200:
          description: Список предложений обмена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trade'
        500:
          $ref: '#/components/responses/ServerError'
  /trades/{trade_id}:
    get:
      tags: [Trades]
      summary: Получить детали предложения обмена
      description: |
        Возвращает полную информацию о конкретном предложении обмена.
      parameters:
        - in: path
          name: trade_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Детали предложения обмена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        404:
          description: Предложение обмена не найдено
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Предложение обмена не найдено'
        500:
          $ref: '#/components/responses/ServerError'
  /trades/initiate:
    post:
      tags: [Trades]
      summary: Создать новое предложение обмена
      description: |
        Создает новое предложение обмена.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                offeredCardId:
                  type: integer
                  example: 401
                  description: ID карты, которую предлагает текущий пользователь
                requestedCardId:
                  type: array
                  items:
                    type: integer
                  description: ID карт, на которые пользователь готов совершить "быстрый обмен"
                  example: [100, 200]
              required: [receivingUserId, requestedCardId, offeredCardId]
      responses:
        201:
          description: Предложение обмена успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        400:
          description: Невозможно создать предложение обмена
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Недостаточно карт для обмена'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /trades/my:
    get:
      tags: [Trades]
      summary: Получение моих обменов
      description: |
        Возвращает список всех обменов текущего пользователя с их статусами.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: integer
            example: 123
            description: 'ID текущего пользователя'
      responses:
        200:
          description: Список моих обменов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trade'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /trades/{trade_id}/accept:
    post:
      tags: [Trades]
      summary: Принять предложение обмена
      description: |
        Принимает предложение обмена, если текущий пользователь - получатель обмена.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: trade_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Обмен успешно принят
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Обмен успешно завершен'

        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          description: Нет прав для принятия этого обмена
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Вы не можете принять этот обмен'
        500:
          $ref: '#/components/responses/ServerError'
  /trades/{trade_id}/reject:
    post:
      tags: [Trades]
      summary: Отклонить предложение обмена
      description: |
        Отклоняет предложение обмена, если текущий пользователь - получатель обмена.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: trade_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Обмен успешно отклонен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Обмен отклонен'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /trades/{trade_id}/cancel:
    post:
      tags: [Trades]
      summary: Отменить предложение обмена
      description: |
        Отменяет предложение обмена, если текущий пользователь является инициатором.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: trade_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Обмен успешно отменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Обмен отменен'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    Achievement:
      type: object
      properties:
        achievement_ID:
          type: integer
          example: 101
          description: 'Уникальный идентификатор достижения'
        name:
          type: string
          example: 'Мастер коллекционирования'
          description: 'Название достижения'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/achievements/master.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Соберите 10 полных коллекций'
          description: 'Описание достижения'
        isUnlocked:
          type: boolean
          example: false
          description: 'Статус получения'
      required:
        - achievement_ID
        - name
        - imageURL
        - description
        - isUnlocked
    Notification:
      type: object
      properties:
        notification_ID:
          type: integer
          example: 201
          description: 'Уникальный идентификатор уведомления'
        user:
          $ref: '#/components/schemas/User'
          description: 'Пользователь, которому отправлено уведомление'
        message:
          type: string
          example: 'Ваша карта была успешно обменяна'
          description: 'Текст уведомления'
        links:
          type: array
          items:
            type: string
            format: url
          example: ['https://example.com/exchange-details']
          description: 'Ссылка на подробности обмена'
        notificationDateTime:
          type: string
          format: date-time
          example: '2024-05-20T15:30:00Z'
          description: 'Дата и время уведомления'
      required:
        - notification_ID
        - user
        - message
        - notificationDateTime
...
    Card:
      type: object
      properties:
        card_ID:
          type: integer
          example: 501
          description: 'Уникальный идентификатор карты'
        name:
          type: string
          example: 'Ледяной феникс'
          description: 'Название карты'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/cards/pheonix.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Ледяной Феникс — это мифическое существо, воплощающее силу зимы и вечного обновления. Его перья сверкают как морозный утренний иней, а глаза светятся холодным синим светом.'
          description: 'Описание карты'
        rarity:
          type: string
          enum: ['Обычная', 'Редкая', 'Эпическая', 'Легендарная', 'Уникальная']
          example: 'Редкая'
          description: 'Редкость карты'
        min_price:
          type: integer
          example: 50
          description: 'Минимальная цена - используется для разбора карточки'
        theme:
          type: string
          nullable: true
          example: 'Мифическое существо'
          description: 'Тематика, на которую была сгенерирвана карта (только для уникальных карт)'
        isGenerated:
          type: boolean
          example: false
          description: 'Флаг показывающий сгенерированная ли карта (true для уникальных, false для остальных)'
      required:
        - card_ID
        - name
        - imageURL
        - rarity
        - min_price
        - isGenerated
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
          description: 'Уникальный идентификатор'
        username:
          type: string
          example: 'Коллекционер_123'
          description: 'Имя пользователя'
        email:
          type: string
          format: email
          example: 'user@example.com'
          description: 'Электронная почта'
        password:
          type: string
          format: password
          writeOnly: true
          description: 'Пароль'
        favoriteCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Избранные карты'
        onChange:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты на обмен'
        inventoryCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты в инвентаре'
        balance:
          type: integer
          example: 2500
          description: 'Баланс'
        avatar_url:
          type: string
          format: url
          example: 'https://example.com/avatars/user1.png'
          description: 'Аватар пользователя'
        achievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
          description: 'Достижения'
        favoriteAchievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
          description: 'Избранные достижения'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
          description: 'Уведомления'
      required:
        - user_id
        - username
        - email
        - password
        - balance
        - inventoryCards
        - achievements
...
    Trade:
      type: object
      properties:
        trade_ID:
          type: integer
          example: 701
          description: 'Уникальный идентификатор обмена'
        offeringUser:
          $ref: '#/components/schemas/User'
          description: 'Пользователь, предлагающий обмен'
        offeringCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Предлагаемые карты'
        receivingUser:
          $ref: '#/components/schemas/User'
          description: 'Пользователь, получающий предложение'
        receivingCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Запрашиваемые карты'
        isConfirmed:
          type: boolean
          example: false
          description: 'Статус подтверждения'
        tradeDateTime:
          type: string
          format: date-time
          example: '2024-05-20T18:00:00Z'
          description: 'Дата и время обмена'
      required:
        - trade_ID
        - offeringUser
        - offeringCards
        - receivingUser
        - receivingCards
        - isConfirmed
        - tradeDateTime
...
  responses:
...
    Unauthorized:
      description: Ошибка аутентификации
    ServerError:
      description: Что-то пошло не так
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Что-то пошло не так, попробуйте позже'
              code:
                type: string
                example: 'INTERNAL_SERVER_ERROR'
              timestamp:
                type: string
                format: date-time
                example: '2024-05-20T12:00:00Z'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT